Scrolling list with highlighting and scroll arrows
This change will provide a highlighted selection for the stories list. Arrow indicators are added to show when stories are out of view and can be scrolled to. To tighten up the UI, moved the copyright string under the banner (and changed to a format that is internationally recognized. This provided room for text showing the number of stories available.